+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Changes since v2.12
 - Application attachment fields now accept multiple attachments. (#2122)
 - It's now possible to add a text to the login page after the login button using extra translations (:t.login/intro2) (#2214)
 - Indicate which items are in shopping cart by changing add button to remove (#2228)
+- Applicants now receive an email when submitting an application. (#2234)
 
 ## v2.12 "Merituulentie" 2020-05-04